refactor applyTrackingParams as a lib to pass init tests
fix typescript
