+++ templates/skeleton/app/lib/search.ts
@@ -1,0 +1,29 @@
+import type {
+  PredictiveQueryFragment,
+  SearchProductFragment,
+  PredictiveProductFragment,
+  PredictiveCollectionFragment,
+  PredictivePageFragment,
+  PredictiveArticleFragment,
+} from 'storefrontapi.generated';
+
+export function applyTrackingParams(
+  resource:
+    | PredictiveQueryFragment
+    | SearchProductFragment
+    | PredictiveProductFragment
+    | PredictiveCollectionFragment
+    | PredictiveArticleFragment
+    | PredictivePageFragment,
+  params?: string,
+) {
+  if (params) {
+    return resource?.trackingParameters
+      ? `?${params}&${resource.trackingParameters}`
+      : `?${params}`;
+  } else {
+    return resource?.trackingParameters
+      ? `?${resource.trackingParameters}`
+      : '';
+  }
+}
